requirements.txt
asgiref==3.7.2
Django==4.0.10
psycopg2-binary==2.9.9
sqlparse==0.4.4
djangorestframework
markdown
django-filter
drf-yasg